requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-core
langchain-google-genai
langgraph
sqlite3-binary
